+++ document.R
@@ -1,5 +1,0 @@
-if(file.exists("DESCRIPTION")) unlink("docs", recursive = TRUE)
-codemetar::write_codemeta()
-knitr::knit("README.Rmd")
-pkgdown::build_site()
-devtools::check()